HER_SHIELD/HerShield_Project1/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
twilio==8.10.0
google-generativeai==0.3.2 